requirements.txt
streamlit
streamlit-gsheets
google-generativeai
pandas
st-gsheets-connection